+++ packages/room-shell/src/data-sources/index.ts
@@ -1,4 +1,0 @@
-export * from './FileDataSourceCard';
-export * from './FileDataSourcesPanel';
-export * from './TableCard';
-export * from './TablesListPanel';