+++ packages/s3-browser-config/src/index.ts
@@ -2,5 +2,7 @@
  * {@include ../README.md}
  * @packageDocumentation
  */
-export * from './S3FileOrDirectory';
-export * from './S3Credentials';
+
+// Values also export their corresponding types automatically (Zod pattern)
+export {S3FileOrDirectory} from './S3FileOrDirectory';
+export {S3Config, S3Credentials} from './S3Credentials';