+++ packages/sql-editor/src/constants/duckdb-dialect.ts
@@ -240,7 +240,7 @@ export const SQL_LANGUAGE_CONFIGURATION: Monaco.languages.IMonarchLanguage = {
     numbers: [
       [/0[xX][0-9a-fA-F]*/, 'number'],
       [/[$][+-]*\d*(\.\d*)?/, 'number'],
-      [/((\d+(\.\d*)?)|(\.\d+))([eE][\-+]?\d+)?/, 'number'],
+      [/((\d+(\.\d*)?)|(\.\d+))([eE][-+]?\d+)?/, 'number'],
     ],
     strings: [
       [/'/, {token: 'string', next: '@string'}],